Refactor most tracked chart
