+++ app/views/dashboard/_earnings_overview.html.haml
@@ -39,17 +39,17 @@
       - c.with(:body) { content_tag(:canvas, nil, id: "chart_number_access_by_main_services") }
       - c.with(:hint) { t("hint.number_access_by_main_services") }
 
-  -# .col-lg-6
-  -#   - name = t("welcomes.most_service_tracked_by_periodic")
-  -#   = render( ChartComponent.new(name: name, filterable: true) ) do |c|
-  -#     - c.with(:addon) do
-  -#       %select.period-filter{ data: { resourcepath: "service-tracked", formater: "mostServiceTracked", canvasid: "chart_most_service_tracked_periodically" }}
-  -#         %option{ value: "month" }= t("period.monthly")
-  -#         %option{ value: "quarter" }= t("period.quarterly")
-  -#         %option{ value: "year" }= t("period.yearly")
-  -#     - c.with(:hint) { t("hint.most_service_tracked_by_periodic") }
-  -#     - c.with(:body) do
-  -#       %canvas#chart_most_service_tracked_periodically
+  .col-lg-6
+    - name = t("welcomes.most_service_tracked_by_periodic")
+    = render( ChartComponent.new(name: name, filterable: true) ) do |c|
+      - c.with(:addon) do
+        %select.period-filter{ data: { resourcepath: "service-tracked", formater: "mostServiceTracked", canvasid: "chart_most_service_tracked_periodically" }}
+          %option{ value: "month" }= t("period.monthly")
+          %option{ value: "quarter" }= t("period.quarterly")
+          %option{ value: "year" }= t("period.yearly")
+      - c.with(:hint) { t("hint.most_service_tracked_by_periodic") }
+      - c.with(:body) do
+        %canvas#chart_most_service_tracked_periodically
 
   -# .col-lg-6
   -#   = render( ChartComponent.new(name: t("welcomes.ticket_tracking_by_gender")) ) do |c|
